fix(multi-backend): use `react-dnd-touch-backend`'s own Typescript definitions
You may find on recent versions of the backend you need to write `TouchBackend({ ...opts }) as BackendFactory` to avoid the under-specification that TouchBackend(arg) could produce an actual backend if arg is a DragDropManager, so the output type is `TouchBackend | BackendFactory`.

Not a breaking change because those typings were aspirational only, meant as a stopgap re-export so users wouldn't have to wrangle with ambient definitions, not part of multi-backend's public API. The same goes for `dnd-multi-backend`.